Issue #24720 remove Template
