+++ dev/io.openliberty.data.internal/src/io/openliberty/data/internal/cdi/BeanDefiningAnnotationMetadata.java
@@ -19,7 +19,6 @@ import org.osgi.service.component.annotations.Component;
 import org.osgi.service.component.annotations.ConfigurationPolicy;
 
 import io.openliberty.cdi.spi.CDIExtensionMetadata;
-import jakarta.data.Entities;
 import jakarta.data.repository.Repository;
 
 /**
@@ -30,7 +29,7 @@ import jakarta.data.repository.Repository;
            service = { CDIExtensionMetadata.class, BeanDefiningAnnotationMetadata.class },
            immediate = true)
 public class BeanDefiningAnnotationMetadata implements CDIExtensionMetadata {
-    private static final Set<Class<? extends Annotation>> beanDefiningAnnos = Set.of(Entities.class, Repository.class);
+    private static final Set<Class<? extends Annotation>> beanDefiningAnnos = Set.of(Repository.class);
 
     @Override
     public Set<Class<? extends Annotation>> getBeanDefiningAnnotationClasses() {
