+++ dev/io.openliberty.data.internal_fat/bnd.bnd
@@ -1,5 +1,5 @@
 #*******************************************************************************
-# Copyright (c) 2022 IBM Corporation and others.
+# Copyright (c) 2022,2023 IBM Corporation and others.
 # All rights reserved. This program and the accompanying materials
 # are made available under the terms of the Eclipse Public License 2.0
 # which accompanies this distribution, and is available at
@@ -16,8 +16,7 @@ bVersion=1.0
 src: \
   fat/src,\
   test-applications/DataTestApp/src,\
-  test-applications/ProviderTestApp/src,\
-  test-applications/TemplateTestApp/src
+  test-applications/ProviderTestApp/src
 
 # TODO 17
 javac.source: 11
