--- dev/io.openliberty.data.internal_fat/publish/servers/io.openliberty.data.internal.template.fat/bootstrap.properties
@@ -1,14 +1,0 @@
-###############################################################################
-# Copyright (c) 2022 IBM Corporation and others.
-# All rights reserved. This program and the accompanying materials
-# are made available under the terms of the Eclipse Public License 2.0
-# which accompanies this distribution, and is available at
-# http://www.eclipse.org/legal/epl-2.0/
-# 
-# SPDX-License-Identifier: EPL-2.0
-#
-# Contributors:
-#     IBM Corporation - initial API and implementation
-###############################################################################
-bootstrap.include=../testports.properties
-com.ibm.ws.logging.trace.specification=*=info:data=all:persistenceService=all:RRA=all
