--- dev/io.openliberty.data.internal_fat/publish/servers/io.openliberty.data.internal.template.fat/server.xml
@@ -1,42 +1,0 @@
-<!--
-    Copyright (c) 2022 IBM Corporation and others.
-    All rights reserved. This program and the accompanying materials
-    are made available under the terms of the Eclipse Public License 2.0
-    which accompanies this distribution, and is available at
-    http://www.eclipse.org/legal/epl-2.0/
-    
-    SPDX-License-Identifier: EPL-2.0
-   
-    Contributors:
-        IBM Corporation - initial API and implementation
- -->
-<server>
-
-  <featureManager>
-    <feature>componenttest-2.0</feature>
-    <feature>data-1.0</feature>
-    <feature>jdbc-4.2</feature>
-    <feature>jndi-1.0</feature>
-    <feature>servlet-6.0</feature>
-  </featureManager>
-
-  <include location="../fatTestPorts.xml"/>
-
-  <application location="TemplateTestApp.war"/>
-
-  <databaseStore id="defaultDatabaseStore" createTables="true" tablePrefix="DATA" dataSourceRef="DerbyDataSource">
-    <authData user="dbuser1" password="dbpwd1"/>
-  </databaseStore>
-
-  <dataSource id="DerbyDataSource">
-    <jdbcDriver libraryRef="DerbyLib"/>
-    <properties.derby.embedded createDatabase="create" databaseName="memory:testdb"/>
-  </dataSource>
-
-  <library id="DerbyLib">
-    <file name="${shared.resource.dir}/derby/derby.jar"/>
-  </library>  
-
-  <javaPermission codeBase="${shared.resource.dir}/derby/derby.jar" className="java.security.AllPermission"/>
-
-</server>
