+++ dev/io.openliberty.data.internal_fat/test-applications/DataTestApp/src/test/jakarta/data/web/House.java
@@ -1,17 +1,36 @@
 /*******************************************************************************
- * Copyright (c) 2022 IBM Corporation and others.
+ * Copyright (c) 2022,2023 IBM Corporation and others.
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License 2.0
  * which accompanies this distribution, and is available at
  * http://www.eclipse.org/legal/epl-2.0/
- * 
+ *
  * SPDX-License-Identifier: EPL-2.0
  *
  * Contributors:
- * IBM Corporation - initial API and implementation
+ *     IBM Corporation - initial API and implementation
  *******************************************************************************/
+package test.jakarta.data.web;
+
+import java.time.Year;
+
 /**
- * @version 1.0
+ * A simple entity without annotations.
  */
-@org.osgi.annotation.versioning.Version("1.0")
-package jakarta.data;
+public class House {
+    public int area;
+
+    public Garage garage;
+
+    public Kitchen kitchen;
+
+    public float lotSize;
+
+    public int numBedrooms;
+
+    public String parcelId;
+
+    public float purchasePrice;
+
+    public Year sold;
+}
