+++ dev/io.openliberty.data.internal_fat/test-applications/DataTestApp/src/test/jakarta/data/web/Houses.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2022,2023 IBM Corporation and others.
+ * Copyright (c) 2023 IBM Corporation and others.
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License 2.0
  * which accompanies this distribution, and is available at
@@ -10,19 +10,25 @@
  * Contributors:
  *     IBM Corporation - initial API and implementation
  *******************************************************************************/
-package test.jakarta.data.template.web;
+package test.jakarta.data.web;
+
+import java.util.List;
+
+import jakarta.data.repository.Repository;
 
 /**
- * A simple entity, unannotated.
+ * Repository for operations on the unannotated House entity,
+ * which has multiple levels of unannotated embeddables.
  */
-public class Vehicle {
-    public String make;
+@Repository
+public interface Houses {
+    long deleteAll();
 
-    public String model;
+    long deleteById(String parcel);
 
-    public int numSeats;
+    boolean existsById(String parcel);
 
-    public float price;
+    House findById(String parcel);
 
-    public String vinId;
+    List<House> save(House... h);
 }
