+++ dev/io.openliberty.data.internal_fat/test-applications/DataTestApp/src/test/jakarta/data/web/Kitchen.java
@@ -10,19 +10,14 @@
  * Contributors:
  *     IBM Corporation - initial API and implementation
  *******************************************************************************/
-package test.jakarta.data.template.web;
+package test.jakarta.data.web;
 
 /**
- * An embeddable without annotations at depth 1, with a field that is another embeddable type.
+ * A simple embeddable without annotations.
  */
-public class Garage {
-    public static enum Type {
-        Attached, Detached, TuckUnder
-    };
+public class Kitchen {
 
-    public int area;
+    public int length;
 
-    public GarageDoor door;
-
-    public Type type;
+    public int width;
 }
