+++ dev/io.openliberty.data.internal_fat/test-applications/ProviderTestApp/src/test/jakarta/data/inmemory/web/ProviderTestServlet.java
@@ -1,10 +1,10 @@
 /*******************************************************************************
- * Copyright (c) 2022 IBM Corporation and others.
+ * Copyright (c) 2022,2023 IBM Corporation and others.
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License 2.0
  * which accompanies this distribution, and is available at
  * http://www.eclipse.org/legal/epl-2.0/
- * 
+ *
  * SPDX-License-Identifier: EPL-2.0
  *
  * Contributors:
@@ -33,7 +33,7 @@ public class ProviderTestServlet extends FATServlet {
     Palindromes palindromes;
 
     /**
-     * Uses a template to insert, update, find, and delete entities.
+     * Uses a repository from a mock Jakarta Data provider to insert, update, find, and delete entities.
      */
     @Test
     public void testDataProvider() {
